feat: int to uint for memcpy

+++ include/LCQProblem.ipp
@@ -30,7 +30,7 @@ namespace lcqpOASES {
 			return LCQPOBJECT_NOT_SETUP;
 
 		H = new double[nV*nV];
-		memcpy( H, H_new, nV*nV*sizeof(double) );
+		memcpy( H, H_new, (uint)(nV*nV)*sizeof(double) );
 
 		return SUCCESSFUL_RETURN;
 	}
@@ -48,7 +48,7 @@ namespace lcqpOASES {
 			return INVALID_ARGUMENT;
 
 		g = new double[nV];
-		memcpy( g, g_new, nV*sizeof(double) );
+		memcpy( g, g_new, (uint)nV*sizeof(double) );
 
 		return SUCCESSFUL_RETURN;
 	}
@@ -66,7 +66,7 @@ namespace lcqpOASES {
 
 		if ( lb_new != 0 )
 		{
-			memcpy( lb,lb_new,nV*sizeof(double) );
+			memcpy( lb, lb_new, (uint)nV*sizeof(double) );
 		}
 		else
 		{
@@ -111,7 +111,7 @@ namespace lcqpOASES {
 
 		if ( ub_new != 0 )
 		{
-			memcpy( ub,ub_new,nV*sizeof(double) );
+			memcpy( ub, ub_new, (uint)nV*sizeof(double) );
 		}
 		else
 		{
@@ -155,7 +155,7 @@ namespace lcqpOASES {
 		xk = new double[nV];
 
 		if (_x0 != 0) {
-			memcpy(xk, _x0, nV*sizeof(double));
+			memcpy(xk, _x0, (uint)nV*sizeof(double));
 		} else {
 			for (int i = 0; i < nV; i++)
 				xk[i] = 0;
@@ -163,7 +163,7 @@ namespace lcqpOASES {
 			
 		if (_y0 != 0) {
 			yk = new double[nV + nC + 2*nComp];
-			memcpy(yk, _y0, (nV + nC + 2*nComp)*sizeof(double));
+			memcpy(yk, _y0, (uint)(nV + nC + 2*nComp)*sizeof(double));
 		} else {
 			yk = (double*)0;
 		}
